JWT implementation in progress

+++ api.go
@@ -1,12 +1,14 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
 	"strconv"
 
+	jwt "github.com/golang-jwt/jwt/v5"
 	"github.com/gorilla/mux"
 )
 
@@ -14,6 +16,10 @@ type APIServer struct {
 	listenAddr string
 	store      Storage
 }
+
+type contextKey struct{}
+
+var ClaimsKey = contextKey{}
 
 func NewAPIServer(listenAddr string, store Storage) *APIServer {
 	return &APIServer{
@@ -25,8 +31,8 @@ func NewAPIServer(listenAddr string, store Storage) *APIServer {
 func (s *APIServer) Run() {
 	router := mux.NewRouter()
 	router.HandleFunc("/account", makeHTTPHandleFunc(s.handleAccount))
-	router.HandleFunc("/account/{id}", makeHTTPHandleFunc(s.handleGetAccountByID))
-	router.HandleFunc("/transfer", makeHTTPHandleFunc(s.handleTransfer))
+	router.HandleFunc("/account/{id}", withJWTAuth(makeHTTPHandleFunc(s.handleGetAccountByID)))
+	router.HandleFunc("/transfer", withJWTAuth(makeHTTPHandleFunc(s.handleTransfer)))
 	log.Println("JSON API server running on port: ", s.listenAddr)
 	err := http.ListenAndServe(s.listenAddr, router)
 	if err != nil {
@@ -80,26 +86,57 @@ func (s *APIServer) handleCreateAccount(w http.ResponseWriter, r *http.Request) 
 	if err := s.store.CreateAccount(account); err != nil {
 		return err
 	}
-	return WriteJSON(w, http.StatusOK, account)
+	tokenStr, err := generateJWT(account)
+	if err != nil {
+		return err
+	}
+	return WriteJSON(w, http.StatusOK, tokenStr)
 }
 
 func (s *APIServer) handleDeleteAccount(w http.ResponseWriter, r *http.Request) error {
+	claims, err := getClaimsFromContext(r)
+	if err != nil {
+		return err
+	}
+
+	userIDFromToken := int(claims["accountNumber"].(float64)) // JWT numbers come as float64
+
 	id, err := getID(r)
 	if err != nil {
 		return err
 	}
+
+	// Authorization check
+	if userIDFromToken != id {
+		return fmt.Errorf("forbidden: cannot delete another user's account")
+	}
+
 	if err := s.store.DeleteAccount(id); err != nil {
 		return err
 	}
+
 	return WriteJSON(w, http.StatusOK, map[string]int{"deleted": id})
 }
 
 func (s *APIServer) handleTransfer(w http.ResponseWriter, r *http.Request) error {
+	claims, err := getClaimsFromContext(r)
+	if err != nil {
+		return err
+	}
+	userIDFromToken := int(claims["accountNumber"].(float64))
 	transferReq := new(TransferRequest)
 	if err := json.NewDecoder(r.Body).Decode(transferReq); err != nil {
 		return err
 	}
+	id, err := getID(r)
+	if err != nil {
+		return err
+	}
 	defer r.Body.Close()
+
+	if userIDFromToken != id {
+		return fmt.Errorf("forbidden: cannot transfer from another user's account")
+	}
 
 	return WriteJSON(w, http.StatusOK, transferReq)
 }
@@ -108,6 +145,52 @@ func WriteJSON(w http.ResponseWriter, status int, v any) error {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(status)
 	return json.NewEncoder(w).Encode(v)
+}
+
+func getClaimsFromContext(r *http.Request) (jwt.MapClaims, error) {
+	claims, ok := r.Context().Value(ClaimsKey).(jwt.MapClaims)
+	if !ok {
+		return nil, fmt.Errorf("unauthorized: no claims found")
+	}
+	return claims, nil
+}
+
+func generateJWT(account *Account) (string, error) {
+	secret := GoDotEnvVariable("JWT_SECRET")
+	claims := &jwt.MapClaims{
+		"expiresAt":     15000,
+		"accountNumber": account.Number,
+	}
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	return token.SignedString([]byte(secret))
+}
+
+func withJWTAuth(handlerFunc http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		tokenStr := r.Header.Get("x-jwt-token")
+		token, err := validateJWT(tokenStr)
+		if err != nil {
+			WriteJSON(w, http.StatusForbidden, APIError{Error: "permission denied"})
+			return
+		}
+		if !token.Valid {
+			WriteJSON(w, http.StatusForbidden, APIError{Error: "permission denied"})
+		}
+		if claims, ok := token.Claims.(jwt.MapClaims); ok {
+			// Store claims in request context
+			ctx := context.WithValue(r.Context(), ClaimsKey, claims)
+			// Create new request with updated context
+			r = r.WithContext(ctx)
+		}
+		handlerFunc(w, r)
+	}
+}
+func validateJWT(tokenString string) (*jwt.Token, error) {
+	secret := GoDotEnvVariable("JWT_SECRET")
+	_, err := jwt.Parse(tokenString, func(token *jwt.Token) (any, error) {
+		return []byte(secret), nil
+	}, jwt.WithValidMethods([]string{jwt.SigningMethodHS256.Alg()}))
+	return nil, err
 }
 
 type apiFunc func(http.ResponseWriter, *http.Request) error
